requirements.txt
streamlit==1.40.2
streamlit-option-menu==0.4.0
Pillow==11.0.0